+++ test/macros.jl
@@ -3,7 +3,9 @@
 
 # Check for changes in Julia's expression parsing
 sumexpr = :(sum{x[i,j] * y[i,j], i = 1:N, j = 1:M; i != j})
-@test string(sumexpr) == ":(sum{\$(Expr(:parameters, :((i!=j)))),*(x[i,j],y[i,j]),i = 1:N,j = 1:M})"
+if VERSION >= v"0.3.0-"
+    @test string(sumexpr) == ":(sum{Expr(:parameters, :(i != j)),x[i,j] * y[i,j],i = 1:N,j = 1:M})"
+end
 @test sumexpr.head == :curly
 @test length(sumexpr.args) == 5
 @test sumexpr.args[1] == :sum
@@ -14,7 +16,9 @@ sumexpr = :(sum{x[i,j] * y[i,j], i = 1:N, j = 1:M; i != j})
 
 
 sumexpr = :(sum{x[i,j] * y[i,j], i = 1:N, j = 1:M})
-@test string(sumexpr) == ":(sum{*(x[i,j],y[i,j]),i = 1:N,j = 1:M})"
+if VERSION >= v"0.3.0-"
+    @test string(sumexpr) == ":(sum{x[i,j] * y[i,j],i = 1:N,j = 1:M})"
+end
 @test sumexpr.head == :curly
 @test length(sumexpr.args) == 4
 @test sumexpr.args[1] == :sum