+++ NEWS.md
@@ -1,6 +1,13 @@
 JuMP release notes
 ==================
 
+Version 0.3.2 (February 17, 2014)
+---------------------------------
+
+ * Improved model printing
+   - Preliminary support for IJulia output
+
+
 Version 0.3.1 (January 30, 2014)
 --------------------------------